Split up bottom tab navigation with stack navigation for readability
