+++ assets/pulse.json
@@ -1,0 +1,1 @@
+{"v":"5.1.16","fr":60,"ip":0,"op":120,"w":964,"h":767,"nm":"Practice","ddd":0,"assets":[],"layers":[{"ddd":0,"ind":1,"ty":4,"nm":"Shape Layer 2","sr":1,"ks":{"o":{"a":0,"k":100,"ix":11},"r":{"a":0,"k":0,"ix":10},"p":{"a":0,"k":[486,379.5,0],"ix":2},"a":{"a":0,"k":[-166,38,0],"ix":1},"s":{"a":1,"k":[{"i":{"x":[0.667,0.667,0.667],"y":[1,1,1]},"o":{"x":[0.333,0.333,0.333],"y":[0,0,0]},"n":["0p667_1_0p333_0","0p667_1_0p333_0","0p667_1_0p333_0"],"t":15,"s":[0,0,100],"e":[49,49,100]},{"i":{"x":[0.833,0.833,0.833],"y":[1,1,1]},"o":{"x":[0.333,0.333,0.333],"y":[0,0,0]},"n":["0p833_1_0p333_0","0p833_1_0p333_0","0p833_1_0p333_0"],"t":35,"s":[49,49,100],"e":[100,100,100]},{"i":{"x":[0.833,0.833,0.833],"y":[1,1,1]},"o":{"x":[0.167,0.167,0.167],"y":[0,0,0]},"n":["0p833_1_0p167_0","0p833_1_0p167_0","0p833_1_0p167_0"],"t":50,"s":[100,100,100],"e":[103,103,100]},{"i":{"x":[0.833,0.833,0.833],"y":[1,1,1]},"o":{"x":[0.167,0.167,0.167],"y":[0,0,0]},"n":["0p833_1_0p167_0","0p833_1_0p167_0","0p833_1_0p167_0"],"t":72,"s":[103,103,100],"e":[100,100,100]},{"t":95}],"ix":6}},"ao":0,"shapes":[{"ty":"gr","it":[{"d":1,"ty":"el","s":{"a":0,"k":[246,246],"ix":2},"p":{"a":0,"k":[0,0],"ix":3},"nm":"Ellipse Path 1","mn":"ADBE Vector Shape - Ellipse","hd":false},{"ty":"st","c":{"a":0,"k":[1,0.626989603043,0.039215683937,1],"ix":3},"o":{"a":0,"k":100,"ix":4},"w":{"a":1,"k":[{"i":{"x":[0.667],"y":[1]},"o":{"x":[0.333],"y":[0]},"n":["0p667_1_0p333_0"],"t":35,"s":[259],"e":[16]},{"i":{"x":[0.702],"y":[5.738]},"o":{"x":[0.3],"y":[0]},"n":["0p702_5p738_0p3_0"],"t":50,"s":[16],"e":[16]},{"i":{"x":[0.784],"y":[1]},"o":{"x":[0.428],"y":[0.227]},"n":["0p784_1_0p428_0p227"],"t":95,"s":[16],"e":[0]},{"t":119}],"ix":5},"lc":1,"lj":1,"ml":4,"nm":"Stroke 1","mn":"ADBE Vector Graphic - Stroke","hd":false},{"ty":"tr","p":{"a":0,"k":[-169,37.5],"ix":2},"a":{"a":0,"k":[0,0],"ix":1},"s":{"a":0,"k":[107.813,107.813],"ix":3},"r":{"a":0,"k":0,"ix":6},"o":{"a":0,"k":100,"ix":7},"sk":{"a":0,"k":0,"ix":4},"sa":{"a":0,"k":0,"ix":5},"nm":"Transform"}],"nm":"Ellipse 1","np":3,"cix":2,"ix":1,"mn":"ADBE Vector Group","hd":false}],"ip":0,"op":120,"st":0,"bm":0},{"ddd":0,"ind":2,"ty":4,"nm":"Shape Layer 1","sr":1,"ks":{"o":{"a":0,"k":100,"ix":11},"r":{"a":0,"k":0,"ix":10},"p":{"a":0,"k":[486,379.5,0],"ix":2},"a":{"a":0,"k":[-166,38,0],"ix":1},"s":{"a":1,"k":[{"i":{"x":[0.667,0.667,0.667],"y":[1,1,1]},"o":{"x":[0.333,0.333,0.333],"y":[0,0,0]},"n":["0p667_1_0p333_0","0p667_1_0p333_0","0p667_1_0p333_0"],"t":0,"s":[100,100,100],"e":[110,110,100]},{"i":{"x":[0.667,0.667,0.667],"y":[1,1,1]},"o":{"x":[0.333,0.333,0.333],"y":[0,0,0]},"n":["0p667_1_0p333_0","0p667_1_0p333_0","0p667_1_0p333_0"],"t":15,"s":[110,110,100],"e":[100,100,100]},{"i":{"x":[0,0,0.833],"y":[0.977,0.977,1]},"o":{"x":[0.333,0.333,0.333],"y":[0,0,0]},"n":["0_0p977_0p333_0","0_0p977_0p333_0","0p833_1_0p333_0"],"t":30,"s":[100,100,100],"e":[0,0,100]},{"i":{"x":[0.593,0.593,0.578],"y":[1.092,1.092,2.303]},"o":{"x":[0.259,0.259,0.169],"y":[0.272,0.272,0]},"n":["0p593_1p092_0p259_0p272","0p593_1p092_0p259_0p272","0p578_2p303_0p169_0"],"t":35,"s":[0,0,100],"e":[0,0,100]},{"i":{"x":[0.638,0.638,0.704],"y":[0.971,0.971,-9.725]},"o":{"x":[0.304,0.304,0.349],"y":[-0.049,-0.049,-0.77]},"n":["0p638_0p971_0p304_-0p049","0p638_0p971_0p304_-0p049","0p704_-9p725_0p349_-0p77"],"t":70,"s":[0,0,100],"e":[0,0,100]},{"i":{"x":[0.794,0.794,0.794],"y":[1,1,1]},"o":{"x":[0.715,0.715,0.436],"y":[0.028,0.028,15.161]},"n":["0p794_1_0p715_0p028","0p794_1_0p715_0p028","0p794_1_0p436_15p161"],"t":95,"s":[0,0,100],"e":[100,100,100]},{"t":119}],"ix":6}},"ao":0,"shapes":[{"ty":"gr","it":[{"d":1,"ty":"el","s":{"a":0,"k":[246,246],"ix":2},"p":{"a":0,"k":[0,0],"ix":3},"nm":"Ellipse Path 1","mn":"ADBE Vector Shape - Ellipse","hd":false},{"ty":"st","c":{"a":0,"k":[1,1,1,1],"ix":3},"o":{"a":0,"k":100,"ix":4},"w":{"a":0,"k":0,"ix":5},"lc":1,"lj":1,"ml":4,"nm":"Stroke 1","mn":"ADBE Vector Graphic - Stroke","hd":false},{"ty":"fl","c":{"a":0,"k":[0.060361396521,0.304709106684,0.615686297417,1],"ix":4},"o":{"a":0,"k":100,"ix":5},"r":1,"nm":"Fill 1","mn":"ADBE Vector Graphic - Fill","hd":false},{"ty":"tr","p":{"a":0,"k":[-169,37.5],"ix":2},"a":{"a":0,"k":[0,0],"ix":1},"s":{"a":0,"k":[107.813,107.813],"ix":3},"r":{"a":0,"k":0,"ix":6},"o":{"a":0,"k":100,"ix":7},"sk":{"a":0,"k":0,"ix":4},"sa":{"a":0,"k":0,"ix":5},"nm":"Transform"}],"nm":"Ellipse 1","np":3,"cix":2,"ix":1,"mn":"ADBE Vector Group","hd":false}],"ip":0,"op":120,"st":0,"bm":0}],"markers":[]}
