+++ firebase.json
@@ -5,6 +5,13 @@
       "firebase.json",
       "**/.*",
       "**/node_modules/**"
+    ],
+    "appAssociation": "AUTO",
+    "rewrites": [
+      {
+        "source": "/orbital2022/**",
+        "dynamicLinks": true
+      }
     ]
   }
 }
