+++ loadinganimation.js
@@ -1,0 +1,21 @@
+import React from "react";
+import { StyleSheet, View, Text } from "react-native";
+import LottieView from "lottie-react-native";
+import { useState } from "react";
+export default function SimpleLottie() {
+  return (
+    <View>
+      <LottieView
+        source={require("./assets/pulse.json")}
+        style={styles.animation}
+        autoPlay
+      />
+    </View>
+  );
+}
+const styles = StyleSheet.create({
+  animation: {
+    width: 100,
+    height: 100,
+  },
+});
