+++ package.json
@@ -18,6 +18,8 @@
     "expo": "~45.0.0",
     "expo-application": "~4.1.0",
     "expo-status-bar": "~1.3.0",
+    "firebase": "9.6.11",
+    "lottie-react-native": "^5.1.3",
     "react": "17.0.2",
     "react-dom": "17.0.2",
     "react-native": "0.68.2",
@@ -27,8 +29,7 @@
     "react-native-screens": "~3.11.1",
     "react-native-vector-icons": "^9.1.0",
     "react-native-web": "0.17.7",
-    "webpack-dev-server": "~3.11.0",
-    "firebase": "9.6.11"
+    "webpack-dev-server": "~3.11.0"
   },
   "devDependencies": {
     "@babel/core": "^7.12.9"
